some problems solved in Error handling

+++ controllers/blog-controller.js
@@ -1,5 +1,6 @@
 const mongoose = require("mongoose");
 const { Blog, BlogComments } = require("../models/blog-model");
+const AppError = require("../utils/AppError");
 
 const catchAsync = (fn) => (req, res, next) => {
   fn(req, res, next).catch(next);
@@ -11,8 +12,11 @@ exports.getAllBlogs = catchAsync(async (req, res) => {
   res.status(200).json({ status: "success", data: blogs });
 });
 
-exports.getBlog = catchAsync(async (req, res) => {
+exports.getBlog = catchAsync(async (req, res, next) => {
   const blog = await Blog.findById(req.params.id);
+  if (!blog) {
+    return next(new AppError(`No blog found with ID ${req.params.id} !`, 404));
+  }
   const commentId = blog._id.toString();
   const comments = await BlogComments.find({ blogId: commentId });
   res.status(200).json({ status: "success", data: { blog, comments } });
@@ -23,21 +27,23 @@ exports.addBlog = catchAsync(async (req, res) => {
   res.status(201).json({ status: "success", data: newBlog });
 });
 
-exports.updateBlog = catchAsync(async (req, res) => {
-  const blog = await Blog.FindByIdAndUpdate(req.params.id, req.body, {
+exports.updateBlog = catchAsync(async (req, res, next) => {
+  const blog = await Blog.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
   });
   if (!blog) {
-    console.log("No blog found!!!");
+    return next(new AppError(`No blog found with ID ${req.params.id} !`, 404));
   }
   res.status(201).json({ status: "success", data: blog });
 });
 
-exports.deleteBlog = catchAsync(async (req, res) => {
-  const blog = await Blog.deleteOne({ _id: req.params.id }).then(
-    await BlogComments.deleteMany({ blogId: req.params.id })
-  );
+exports.deleteBlog = catchAsync(async (req, res, next) => {
+  const blog = await Blog.deleteOne({ _id: req.params.id });
+  if (!blog) {
+    return next(new AppError(`No blog found with ID ${req.params.id} !`, 404));
+  }
+  await BlogComments.deleteMany({ blogId: req.params.id });
   res.status(204).json({ status: "success", data: null });
 });
 
@@ -47,21 +53,26 @@ exports.addComment = catchAsync(async (req, res) => {
   res.status(201).json({ status: "success", data: comment });
 });
 
-exports.updateComment = catchAsync(async (req, res) => {
-  console.log(req.body._id);
-  const comment = await BlogComments.FindByIdAndUpdate(
-    "645f8480a032124023444801",
-    { text: "Bu deneme ile mevcut blog içeriği değiştirilmiş oldu" }
-    // {
-    //   new: true,
-    //   runValidators: true,
-    // }
-  );
+exports.updateComment = catchAsync(async (req, res, next) => {
+  const comment = await BlogComments.findByIdAndUpdate(req.body._id, req.body, {
+    new: true,
+    runValidators: true,
+  });
+  if (!comment) {
+    return next(
+      new AppError(`No comment found with ID ${req.body._id} !`, 404)
+    );
+  }
   res.status(201).json({ status: "success", data: comment });
 });
 
-exports.deleteComment = catchAsync(async (req, res) => {
+exports.deleteComment = catchAsync(async (req, res, next) => {
   const comment = await BlogComments.deleteOne({ _id: req.body._id });
+  if (!comment) {
+    return next(
+      new AppError(`No comment found with ID ${req.body._id} !`, 404)
+    );
+  }
   res.status(204).json({ status: "success", data: null });
 });
 // exports.updateComment = async (req, res) => {
